+++ addon/controllers/edit-form.js
@@ -10,10 +10,9 @@ import FlexberryFileControllerMixin from '../mixins/flexberry-file-controller';
 import needSaveCurrentAgregator from '../utils/need-save-current-agregator';
 import getCurrentAgregator from '../utils/get-current-agregator';
 import PaginatedControllerMixin from '../mixins/paginated-controller';
-import ReloadListMixin from '../mixins/reload-list-mixin';
 import SortableControllerMixin from '../mixins/sortable-controller';
 import LimitedControllerMixin from '../mixins/limited-controller';
-import FolvOnEditControllerMixin from '../mixins/flexberry-objectlistview-on-edit-form-controller';
+import FlexberryOlvToolbarMixin from '../mixins/olv-toolbar-controller';
 import FlexberryObjectlistviewHierarchicalControllerMixin from '../mixins/flexberry-objectlistview-hierarchical-controller';
 
 const { getOwner } = Ember;
@@ -53,11 +52,19 @@ FlexberryLookupMixin,
 ErrorableControllerMixin,
 FlexberryFileControllerMixin,
 PaginatedControllerMixin,
-ReloadListMixin,
 SortableControllerMixin,
 LimitedControllerMixin,
-FlexberryObjectlistviewHierarchicalControllerMixin,
-FolvOnEditControllerMixin, {
+FlexberryOlvToolbarMixin,
+FlexberryObjectlistviewHierarchicalControllerMixin, {
+  /**
+    Controller to show colsconfig modal window.
+
+    @property lookupController
+    @type <a href="http://emberjs.com/api/classes/Ember.InjectedProperty.html">Ember.InjectedProperty</a>
+    @default Ember.inject.controller('colsconfig-dialog')
+  */
+  colsconfigController: Ember.inject.controller('colsconfig-dialog'),
+
   /**
     Flag to enable return to agregator's path if possible.
 
@@ -331,13 +338,22 @@ FolvOnEditControllerMixin, {
     },
 
     /**
+      Hook that executes before deleting all records on all pages.
+      Need to be overriden in corresponding application controller.
+    */
+    beforeDeleteAllRecords(modelName, data) {
+      data.cancel = true;
+      Ember.assert(`Please specify 'beforeDeleteAllRecords' action for '${this.componentName}' list compoenent in corresponding controller`);
+    },
+
+    /**
       Sorting list by column.
 
       @method actions.sortByColumn
       @param {Object} column Column for sorting.
     */
-    sortByColumn: function(column) {
-      this._super.apply(this, [column, 'sorting']);
+    sortByColumn: function(column, componentName) {
+      this._super.apply(this, [column, componentName, 'sorting']);
     },
 
     /**
@@ -346,8 +362,8 @@ FolvOnEditControllerMixin, {
       @method actions.addColumnToSorting
       @param {Object} column Column for sorting.
     */
-    addColumnToSorting: function(column) {
-      this._super.apply(this, [column, 'sorting']);
+    addColumnToSorting: function(column, componentName) {
+      this._super.apply(this, [column, componentName, 'sorting']);
     },
   },
 
@@ -365,68 +381,59 @@ FolvOnEditControllerMixin, {
     this.onSaveActionStarted();
     this.get('appState').loading();
 
-    let _this = this;
-
-    let afterSaveModelFunction = () => {
+    const afterSaveModelFunction = () => {
       this.get('appState').success();
-      _this.onSaveActionFulfilled();
+      this.onSaveActionFulfilled();
       if (close) {
         this.get('appState').reset();
-        _this.close(skipTransition);
+        this.close(skipTransition);
       } else if (!skipTransition) {
-        let routeName = _this.get('routeName');
+        const routeName = this.get('routeName');
         if (routeName.indexOf('.new') > 0) {
-          let qpars = {};
-          let queryParams = _this.get('queryParams');
-          queryParams.forEach(function(item, i, params) {
+          const qpars = {};
+          const queryParams = this.get('queryParams');
+          queryParams.forEach(function(item) {
             qpars[item] = this.get(item);
-          }, _this);
+          }, this);
           let transitionQuery = {};
           transitionQuery.queryParams = qpars;
           transitionQuery.queryParams.recordAdded = true;
-          _this.transitionToRoute(routeName.slice(0, -4), _this.get('model'), transitionQuery);
+          const parentParameters = {
+            parentRoute: this.get('parentRoute'),
+            parentRouteRecordId: this.get('parentRouteRecordId')
+          };
+          transitionQuery.queryParams.parentParameters = parentParameters;
+          this.transitionToRoute(routeName.slice(0, -4), this.get('model'), transitionQuery);
         }
       }
     };
 
-    let savePromise = this.get('model').save().then((model) => {
-      let agragatorModel = getCurrentAgregator.call(this);
-      if (needSaveCurrentAgregator.call(this, agragatorModel)) {
-        return this._saveHasManyRelationships(model).then((result) => {
-          if (result && Ember.isArray(result)) {
-            let arrayWrapper = Ember.A();
-            arrayWrapper.addObjects(result);
-            let errors = arrayWrapper.filterBy('state', 'rejected');
-            return errors.length > 0 ? Ember.RSVP.reject(errors) : agragatorModel.save().then(afterSaveModelFunction);
-          } else {
-            return agragatorModel.save().then(afterSaveModelFunction);
-          }
-        });
-      } else {
-        return this._saveHasManyRelationships(model).then((result) => {
-          if (result && Ember.isArray(result)) {
-            let arrayWrapper = Ember.A();
-            arrayWrapper.addObjects(result);
-            let errors = arrayWrapper.filterBy('state', 'rejected');
-            if (errors.length > 0) {
-              return Ember.RSVP.reject(errors);
-            } else {
-              afterSaveModelFunction();
-            }
-          } else {
-            afterSaveModelFunction();
-          }
-        });
-      }
-    }).catch((errorData) => {
+    let savePromise;
+    const model = this.get('model');
+
+    // This is possible when using offline mode.
+    const agragatorModel = getCurrentAgregator.call(this);
+    if (needSaveCurrentAgregator.call(this, agragatorModel)) {
+      savePromise = this._saveHasManyRelationships(model).then((result) => {
+        const errors = Ember.A(result || []).filterBy('state', 'rejected');
+        if (!Ember.isEmpty(errors)) {
+          return Ember.RSVP.reject(errors);
+        }
+
+        return agragatorModel.save();
+      });
+    } else {
+      const unsavedModels = this._getModelWithHasMany(model).filterBy('hasDirtyAttributes');
+      savePromise = unsavedModels.length > 1 ? this.get('store').batchUpdate(unsavedModels) : model.save();
+    }
+
+    return savePromise.then(afterSaveModelFunction).catch((errorData) => {
       this.get('appState').error();
       this.onSaveActionRejected(errorData);
       return Ember.RSVP.reject(errorData);
     }).finally((data) => {
       this.onSaveActionAlways(data);
     });
-
-    return savePromise;
   },
 
   /**
@@ -785,6 +792,29 @@ FolvOnEditControllerMixin, {
     });
 
     return Ember.RSVP.allSettled(promises);
+  },
+
+  /**
+    Returns an array with the model and all its `hasMany` relationships, obtained recursively, for each model.
+
+    @method _getModelWithHasMany
+    @param {DS.Model} model The object model.
+    @return {Ember.NativeArray} An array with the model and all its `hasMany` relationships.
+  */
+  _getModelWithHasMany(model) {
+    const models = Ember.A([model]);
+
+    model.eachRelationship((name, desc) => {
+      if (desc.kind === 'hasMany') {
+        const hasMany = model.get(name);
+        models.addObjects(hasMany);
+        hasMany.map(this._getModelWithHasMany, this).forEach((hasMany) => {
+          models.addObjects(hasMany);
+        });
+      }
+    });
+
+    return models;
   },
 
   /**