+++ tests/dummy/app/controllers/components-examples/flexberry-objectlistview/settings-example.js
@@ -361,7 +361,7 @@ export default ListFormController.extend({
     '  sortByColumn=(action \"sortByColumn\")<br>' +
     '  addColumnToSorting=(action \"addColumnToSorting\")<br>' +
     '  _availableHierarchicalMode=availableHierarchicalMode<br>' +
-    '  _availableCollExpandMode=availableCollExpandMode<br>' +
+    '  availableCollExpandMode=availableCollExpandMode<br>' +
     '  pages=pages<br>' +
     '  perPageValue=perPageValue<br>' +
     '  perPageValues=perPageValues<br>' +